chore: Extract common cloud sign-in hook
https://harperdb.atlassian.net/browse/STUDIO-588
