+++ src/features/instance/operations/schemas/signInSchema.ts
@@ -1,8 +1,14 @@
+import { zodRequireEmail } from '@/lib/zod/email';
 import { zodRequirePassword } from '@/lib/zod/password';
 import { zodRequireUsername } from '@/lib/zod/username';
 import { z } from 'zod';
 
-export const SignInSchema = z.object({
+export const UsernameSignInSchema = z.object({
 	username: zodRequireUsername,
+	password: zodRequirePassword,
+});
+
+export const EmailSignInSchema = z.object({
+	email: zodRequireEmail,
 	password: zodRequirePassword,
 });
